requirements.txt
aiogram>=3.10,<4.0
yandexcloud>=0.300
grpcio>=1.60
apscheduler>=3.10,<4.0
pydantic-settings>=2.0
python-dotenv>=1.0
